+++ CHANGELOG.md
@@ -1,7 +1,20 @@
-## [0.0.11]
-#### Jan 10, 2022
+## [0.0.14]
+#### February 15, 2022
+#### COMMIT NO : 
+- GtPdf Widget changes for the pdf table issue 
+- GtDropdownSearchWidget changes to handle the dropdown result overlay view context (new input added to provide context type for widget)
+- GtTextFormField widget input options added for `onEditingComplete` function and `inputFormatters` to provide the custom regex checks for TextFormField
 
+## [0.0.13]
+#### February 01, 2022
+#### COMMIT NO : cc283c5
+- GtDropdownSearch widget issues resolved
+
+## [0.0.12]
+#### Jan 10, 2022
+#### COMMIT NO : 7b0e98b
 - Documentation Improvment 
+
 
 ## [0.0.11]
 #### Dec 30, 2021
